+++ rclcpp_cascade_lifecycle/include/rclcpp_cascade_lifecycle/rclcpp_cascade_lifecycle.hpp
@@ -18,6 +18,7 @@
 #include <set>
 #include <map>
 #include <string>
+#include <list>
 
 #include "rclcpp/node_options.hpp"
 #include "rclcpp_lifecycle/lifecycle_node.hpp"
@@ -40,6 +41,7 @@ namespace rclcpp_cascade_lifecycle
 class CascadeLifecycleNode : public rclcpp_lifecycle::LifecycleNode
 {
 public:
+  RCLCPP_SMART_PTR_DEFINITIONS(CascadeLifecycleNode)
   /// Create a new lifecycle node with the specified name.
   /**
    * \param[in] node_name Name of the node.
@@ -101,16 +103,21 @@ private:
   rclcpp::Subscription<cascade_lifecycle_msgs::msg::State>::SharedPtr states_sub_;
 
   rclcpp::TimerBase::SharedPtr timer_;
+  rclcpp::TimerBase::SharedPtr timer_responses_;
 
   std::set<std::string> activators_;
   std::set<std::string> activations_;
   std::map<std::string, uint8_t> activators_state_;
+  std::list<cascade_lifecycle_msgs::msg::Activation> op_pending_;
 
   void activations_callback(const cascade_lifecycle_msgs::msg::Activation::SharedPtr msg);
   void states_callback(const cascade_lifecycle_msgs::msg::State::SharedPtr msg);
   void update_state();
 
   void timer_callback();
+  void timer_responses_callback();
+
+  bool allow_duplicate_names_ {true};
 };
 
 }  // namespace rclcpp_cascade_lifecycle