+++ rclcpp_cascade_lifecycle/CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package rclcpp_cascade_lifecycle
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.2 (2024-10-12)
+------------------
+* Explicit removal of node from executor before destroy
+* Allow to have pubs/sub with standard QoS
+* Contributors: Francisco Martín Rico, Juan Carlos Manzanares Serrano
+
 2.0.0 (2024-04-04)
 ------------------
 * Fix clear_activation